ModulePolicyValidator to own file
